requirements.txt
flask
boto3==1.24.27
botocore==1.27.31
gunicorn
s3transfer==0.6.0
python-dotenv
